+++ api/cache/clear.js
@@ -10,7 +10,7 @@
  *
  * @endpoint POST /api/cache/clear
  *
- * @version 4.1.6
+ * @version 5.0.0
  * @author  Shinei Nouzen
  * @license MIT
  * ======= • ======= • ======= • ======= • =======• =======